Fix partially delete events
The index service provides two delete methods: one that requires just
the media package ID and another that requires both the Event object and
a retract workflow ID. The method that takes only the ID removes the
event from the asset manager, but fails to delete it from the search
index.

As a result, the video remains accessible, but Opencast is unaware of
its existence. This update addresses the issue by utilizing the method
that includes the Event object and workflow ID.

Signed-off-by: Gregor Eichelberger <gregor.eichelberger@tuwien.ac.at>